requirements.txt
slack-bolt==1.18.0
supabase==1.0.3
python-dotenv==1.0.0
requests==2.31.0
httpx==0.23.3
websocket-client==1.6.1 